# Conflicts:
#	src/main/java/com/example/bookstore/controller/MainController.java
#	src/main/java/com/example/bookstore/entity/BookEntity.java
#	src/main/java/com/example/bookstore/model/Book.java
#	src/main/resources/data.sql
#	src/main/resources/schema.sql
#	src/main/resources/templates/book-details-page.html
#	src/main/resources/templates/main-page.html

+++ src/main/java/com/example/bookstore/controller/BookRestController.java
@@ -16,6 +16,7 @@ import java.util.List;
 @RequestMapping("/api")
 @RequiredArgsConstructor
 public class BookRestController {
+
     private final BookService bookService;
 
     @GetMapping("/books")
@@ -25,6 +26,6 @@ public class BookRestController {
 
     @GetMapping("/books/{bookId}")
     public ResponseEntity<Book> getBookById(@PathVariable int bookId) {
-        return bookService.getBookById(bookId).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        return bookService.getBookById((long) bookId).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 }
